% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bill_crosscompare.R
\name{bill_cross_check}
\alias{bill_cross_check}
\title{Check two tables with unique matching ids and generate reports
on duplicated ids and repeated columns when applicable..}
\usage{
bill_cross_check(dt1 = NULL, dt2 = NULL, id = NULL, chk = NULL)
}
\arguments{
\item{dt1}{A table.}

\item{dt2}{A table.}

\item{id}{A column name or a vector of column names}

\item{chk}{A column name}
}
\value{
a data.table
}
\description{
\code{bill_cross_check} returns a merged data.table showing
information regarding ids and repeated columns.
}
\examples{

bill_cross_check(data.frame(col1=c(1,2,3,3), col2=c('a','b','c','c'),
col3=c('-','=','+','-')),data.frame(col1=c(1,2,3), col2=c('a','b','c'),
col3=c('-','=','+')), id=c('col1','col2'), chk='col3')

}
