% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bill_crosscompare.R
\name{bill_cross_count}
\alias{bill_cross_count}
\title{Summarize times being purchased for individual items'
 from multiple grocery shopping lists.}
\usage{
bill_cross_count(
  ldt = list(),
  id = NULL,
  gp = NULL,
  type = "count",
  condstr = ""
)
}
\arguments{
\item{ldt}{A list of grocery tables in data.frame format.}

\item{id}{A column name}

\item{gp}{A column name or a vector of column names}

\item{type}{A string in one of the following:
\itemize{
  \item \code{count} for number of rows in each bill
  \item \code{cond} for conditional counting
  \item \code{condwt} for conditional counting with total in parenthesis
}}

\item{condstr}{A string for conditional counting}
}
\value{
a data.table
}
\description{
\code{bill_cross_count} returns a data.table showing how many times each
items being listed on individual bills.
}
\examples{

bill_cross_count(list(cbind(col1=c('a','b','c'),col2=c(1,2,3)),
cbind(col1=c('d','c','d'),col2=c(4,5,6))), id='col1')

}
