% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_color.R
\name{color_test_palette}
\alias{color_test_palette}
\title{Test a color vector}
\usage{
color_test_palette(colvect = c(), type = "line")
}
\arguments{
\item{colvect}{A vector returned by \code{color_set_palette()}}

\item{type}{One of
\itemize{
  \item 'line' for using the color vector on a line plot (Default)
  \item 'box' for using the color vector on a box plot.
}}
}
\value{
A plot
}
\description{
Create a figure using the assigned color vector
}
\examples{
color_test_palette( setNames(c(1,2,3,4), c('apple','orange','avocado','lime') ))

}
